# game/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is game server.",
        "title": "Game API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/api/v1/game",
    "paths": {
        "/api/v1/game/getGame/{id}": {
            "get": {
                "description": "Get game data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Get game data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Game"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/game/getPrices": {
            "get": {
                "description": "Get prices",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Get prices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.PriceResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/game/updateMultiplicator": {
            "post": {
                "description": "Update multiplicator",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Update multiplicator",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MultipUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.OKStruct"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/game/updateScore": {
            "post": {
                "description": "Update score",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Game"
                ],
                "summary": "Update score",
                "parameters": [
                    {
                        "description": "Body",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Score"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.OKStruct"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Game": {
            "type": "object",
            "properties": {
                "gasMining": {
                    "type": "integer"
                },
                "gasStorage": {
                    "type": "integer"
                },
                "ownerId": {
                    "type": "integer"
                },
                "protection": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
        "model.MultipUpdate": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nameType": {
                    "type": "string"
                }
            }
        },
        "model.Score": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
        "server.OKStruct": {
            "type": "object"
        },
        "server.PriceResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
        "server.errorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "errorInfo": {
                    "type": "string"
                }
            }
        }
    }
}